stream: move RangeBuf implementation to a separate file
