+++ quiche/src/lib.rs
@@ -9361,6 +9361,7 @@ pub mod testing {
 
 #[cfg(test)]
 mod tests {
+    use crate::range_buf::RangeBuf;
     use rstest::rstest;
 
     use super::*;
@@ -10021,7 +10022,7 @@ mod tests {
 
         let frames = [frame::Frame::Stream {
             stream_id: 4,
-            data: stream::RangeBuf::from(b"aaaaa", 0, true),
+            data: RangeBuf::from(b"aaaaa", 0, true),
         }];
 
         assert_eq!(
@@ -10086,7 +10087,7 @@ mod tests {
 
         let frames = [frame::Frame::Stream {
             stream_id: 4,
-            data: stream::RangeBuf::from(b"aaaaa", 0, true),
+            data: RangeBuf::from(b"aaaaa", 0, true),
         }];
 
         let len =
@@ -10160,7 +10161,7 @@ mod tests {
 
         let frames = [frame::Frame::Stream {
             stream_id: 4,
-            data: stream::RangeBuf::from(b"aaaaa", 0, true),
+            data: RangeBuf::from(b"aaaaa", 0, true),
         }];
 
         let len =
@@ -10209,7 +10210,7 @@ mod tests {
         // Client send a 1-byte frame that starts from the crypto stream offset
         // limit.
         let frames = [frame::Frame::Crypto {
-            data: stream::RangeBuf::from(b"a", MAX_CRYPTO_STREAM_OFFSET, false),
+            data: RangeBuf::from(b"a", MAX_CRYPTO_STREAM_OFFSET, false),
         }];
 
         let pkt_type = packet::Type::Short;
@@ -10441,7 +10442,7 @@ mod tests {
 
         let frames = [frame::Frame::Stream {
             stream_id: 4,
-            data: stream::RangeBuf::from(b"aaaaa", 0, false),
+            data: RangeBuf::from(b"aaaaa", 0, false),
         }];
 
         let pkt_type = packet::Type::Short;
@@ -10454,7 +10455,7 @@ mod tests {
 
         let frames = [frame::Frame::Stream {
             stream_id: 4,
-            data: stream::RangeBuf::from(b"", 5, true),
+            data: RangeBuf::from(b"", 5, true),
         }];
 
         let pkt_type = packet::Type::Short;
@@ -10467,7 +10468,7 @@ mod tests {
 
         let frames = [frame::Frame::Stream {
             stream_id: 4,
-            data: stream::RangeBuf::from(b"", 15, true),
+            data: RangeBuf::from(b"", 15, true),
         }];
 
         let pkt_type = packet::Type::Short;
@@ -10544,7 +10545,7 @@ mod tests {
 
         let frames = [frame::Frame::Stream {
             stream_id: 4,
-            data: stream::RangeBuf::from(b"hello", 0, false),
+            data: RangeBuf::from(b"hello", 0, false),
         }];
 
         // Client sends stream frame with key update request.
@@ -10678,15 +10679,15 @@ mod tests {
         let frames = [
             frame::Frame::Stream {
                 stream_id: 0,
-                data: stream::RangeBuf::from(b"aaaaaaaaaaaaaaa", 0, false),
+                data: RangeBuf::from(b"aaaaaaaaaaaaaaa", 0, false),
             },
             frame::Frame::Stream {
                 stream_id: 4,
-                data: stream::RangeBuf::from(b"aaaaaaaaaaaaaaa", 0, false),
+                data: RangeBuf::from(b"aaaaaaaaaaaaaaa", 0, false),
             },
             frame::Frame::Stream {
                 stream_id: 8,
-                data: stream::RangeBuf::from(b"a", 0, false),
+                data: RangeBuf::from(b"a", 0, false),
             },
         ];
 
@@ -10710,16 +10711,16 @@ mod tests {
             // One byte less than stream limit.
             frame::Frame::Stream {
                 stream_id: 0,
-                data: stream::RangeBuf::from(b"aaaaaaaaaaaaaa", 0, false),
+                data: RangeBuf::from(b"aaaaaaaaaaaaaa", 0, false),
             },
             // Same stream, but one byte more.
             frame::Frame::Stream {
                 stream_id: 0,
-                data: stream::RangeBuf::from(b"aaaaaaaaaaaaaaa", 0, false),
+                data: RangeBuf::from(b"aaaaaaaaaaaaaaa", 0, false),
             },
             frame::Frame::Stream {
                 stream_id: 8,
-                data: stream::RangeBuf::from(b"aaaaaaaaaaaaaaa", 0, false),
+                data: RangeBuf::from(b"aaaaaaaaaaaaaaa", 0, false),
             },
         ];
 
@@ -10739,11 +10740,11 @@ mod tests {
         let frames = [
             frame::Frame::Stream {
                 stream_id: 0,
-                data: stream::RangeBuf::from(b"aaaaaaaaaaaaaaa", 0, false),
+                data: RangeBuf::from(b"aaaaaaaaaaaaaaa", 0, false),
             },
             frame::Frame::Stream {
                 stream_id: 4,
-                data: stream::RangeBuf::from(b"a", 0, false),
+                data: RangeBuf::from(b"a", 0, false),
             },
         ];
 
@@ -10756,7 +10757,7 @@ mod tests {
 
         let frames = [frame::Frame::Stream {
             stream_id: 4,
-            data: stream::RangeBuf::from(b"a", 1, false),
+            data: RangeBuf::from(b"a", 1, false),
         }];
 
         let len = pipe
@@ -10829,7 +10830,7 @@ mod tests {
 
         let frames = [frame::Frame::Stream {
             stream_id: 4,
-            data: stream::RangeBuf::from(b"aaaaaaaaaaaaaaaa", 0, true),
+            data: RangeBuf::from(b"aaaaaaaaaaaaaaaa", 0, true),
         }];
 
         let pkt_type = packet::Type::Short;
@@ -10850,7 +10851,7 @@ mod tests {
 
         let frames = [frame::Frame::Stream {
             stream_id: 2,
-            data: stream::RangeBuf::from(b"aaaaaaaaaaa", 0, true),
+            data: RangeBuf::from(b"aaaaaaaaaaa", 0, true),
         }];
 
         let pkt_type = packet::Type::Short;
@@ -10871,7 +10872,7 @@ mod tests {
 
         let frames = [frame::Frame::Stream {
             stream_id: 4,
-            data: stream::RangeBuf::from(b"aaaaaaaaa", 0, false),
+            data: RangeBuf::from(b"aaaaaaaaa", 0, false),
         }];
 
         let pkt_type = packet::Type::Short;
@@ -10882,7 +10883,7 @@ mod tests {
 
         let frames = [frame::Frame::Stream {
             stream_id: 4,
-            data: stream::RangeBuf::from(b"a", 9, false),
+            data: RangeBuf::from(b"a", 9, false),
         }];
 
         let len = pipe
@@ -10975,31 +10976,31 @@ mod tests {
         let frames = [
             frame::Frame::Stream {
                 stream_id: 4,
-                data: stream::RangeBuf::from(b"a", 0, false),
+                data: RangeBuf::from(b"a", 0, false),
             },
             frame::Frame::Stream {
                 stream_id: 8,
-                data: stream::RangeBuf::from(b"a", 0, false),
+                data: RangeBuf::from(b"a", 0, false),
             },
             frame::Frame::Stream {
                 stream_id: 12,
-                data: stream::RangeBuf::from(b"a", 0, false),
+                data: RangeBuf::from(b"a", 0, false),
             },
             frame::Frame::Stream {
                 stream_id: 16,
-                data: stream::RangeBuf::from(b"a", 0, false),
+                data: RangeBuf::from(b"a", 0, false),
             },
             frame::Frame::Stream {
                 stream_id: 20,
-                data: stream::RangeBuf::from(b"a", 0, false),
+                data: RangeBuf::from(b"a", 0, false),
             },
             frame::Frame::Stream {
                 stream_id: 24,
-                data: stream::RangeBuf::from(b"a", 0, false),
+                data: RangeBuf::from(b"a", 0, false),
             },
             frame::Frame::Stream {
                 stream_id: 28,
-                data: stream::RangeBuf::from(b"a", 0, false),
+                data: RangeBuf::from(b"a", 0, false),
             },
         ];
 
@@ -11047,31 +11048,31 @@ mod tests {
         let frames = [
             frame::Frame::Stream {
                 stream_id: 2,
-                data: stream::RangeBuf::from(b"a", 0, false),
+                data: RangeBuf::from(b"a", 0, false),
             },
             frame::Frame::Stream {
                 stream_id: 6,
-                data: stream::RangeBuf::from(b"a", 0, false),
+                data: RangeBuf::from(b"a", 0, false),
             },
             frame::Frame::Stream {
                 stream_id: 10,
-                data: stream::RangeBuf::from(b"a", 0, false),
+                data: RangeBuf::from(b"a", 0, false),
             },
             frame::Frame::Stream {
                 stream_id: 14,
-                data: stream::RangeBuf::from(b"a", 0, false),
+                data: RangeBuf::from(b"a", 0, false),
             },
             frame::Frame::Stream {
                 stream_id: 18,
-                data: stream::RangeBuf::from(b"a", 0, false),
+                data: RangeBuf::from(b"a", 0, false),
             },
             frame::Frame::Stream {
                 stream_id: 22,
-                data: stream::RangeBuf::from(b"a", 0, false),
+                data: RangeBuf::from(b"a", 0, false),
             },
             frame::Frame::Stream {
                 stream_id: 26,
-                data: stream::RangeBuf::from(b"a", 0, false),
+                data: RangeBuf::from(b"a", 0, false),
             },
         ];
 
@@ -11390,15 +11391,15 @@ mod tests {
         let frames = [
             frame::Frame::Stream {
                 stream_id: 0,
-                data: stream::RangeBuf::from(b"aaaaa", 0, false),
+                data: RangeBuf::from(b"aaaaa", 0, false),
             },
             frame::Frame::Stream {
                 stream_id: 0,
-                data: stream::RangeBuf::from(b"bbbbb", 3, false),
+                data: RangeBuf::from(b"bbbbb", 3, false),
             },
             frame::Frame::Stream {
                 stream_id: 0,
-                data: stream::RangeBuf::from(b"ccccc", 6, false),
+                data: RangeBuf::from(b"ccccc", 6, false),
             },
         ];
 
@@ -11422,15 +11423,15 @@ mod tests {
         let frames = [
             frame::Frame::Stream {
                 stream_id: 0,
-                data: stream::RangeBuf::from(b"aaaaa", 0, false),
+                data: RangeBuf::from(b"aaaaa", 0, false),
             },
             frame::Frame::Stream {
                 stream_id: 0,
-                data: stream::RangeBuf::from(b"ccccc", 6, false),
+                data: RangeBuf::from(b"ccccc", 6, false),
             },
             frame::Frame::Stream {
                 stream_id: 0,
-                data: stream::RangeBuf::from(b"bbbbb", 3, false),
+                data: RangeBuf::from(b"bbbbb", 3, false),
             },
         ];
 
@@ -11589,11 +11590,11 @@ mod tests {
         let frames = [
             frame::Frame::Stream {
                 stream_id: 0,
-                data: stream::RangeBuf::from(b"aaaaaaaaaaaaaaa", 0, false),
+                data: RangeBuf::from(b"aaaaaaaaaaaaaaa", 0, false),
             },
             frame::Frame::Stream {
                 stream_id: 4,
-                data: stream::RangeBuf::from(b"a", 0, false),
+                data: RangeBuf::from(b"a", 0, false),
             },
             frame::Frame::ResetStream {
                 stream_id: 4,
@@ -11602,7 +11603,7 @@ mod tests {
             },
             frame::Frame::Stream {
                 stream_id: 8,
-                data: stream::RangeBuf::from(b"a", 0, false),
+                data: RangeBuf::from(b"a", 0, false),
             },
         ];
 
@@ -11627,7 +11628,7 @@ mod tests {
         let frames = [
             frame::Frame::Stream {
                 stream_id: 4,
-                data: stream::RangeBuf::from(b"a", 0, false),
+                data: RangeBuf::from(b"a", 0, false),
             },
             frame::Frame::ResetStream {
                 stream_id: 4,
@@ -11708,7 +11709,7 @@ mod tests {
         // Send 1-RTT packet #0.
         let frames = [frame::Frame::Stream {
             stream_id: 0,
-            data: stream::RangeBuf::from(b"hello, world", 0, true),
+            data: RangeBuf::from(b"hello, world", 0, true),
         }];
 
         let pkt_type = packet::Type::Short;
@@ -11721,7 +11722,7 @@ mod tests {
         // Send 1-RTT packet #1.
         let frames = [frame::Frame::Stream {
             stream_id: 4,
-            data: stream::RangeBuf::from(b"hello, world", 0, true),
+            data: RangeBuf::from(b"hello, world", 0, true),
         }];
 
         let written =
@@ -12418,7 +12419,7 @@ mod tests {
             iter.next(),
             Some(&frame::Frame::Stream {
                 stream_id: 8,
-                data: stream::RangeBuf::from(b"aaaaa", 0, false),
+                data: RangeBuf::from(b"aaaaa", 0, false),
             })
         );
 
@@ -12431,7 +12432,7 @@ mod tests {
             frames.first(),
             Some(&frame::Frame::Stream {
                 stream_id: 0,
-                data: stream::RangeBuf::from(b"aaaaa", 0, false),
+                data: RangeBuf::from(b"aaaaa", 0, false),
             })
         );
 
@@ -12444,7 +12445,7 @@ mod tests {
             frames.first(),
             Some(&frame::Frame::Stream {
                 stream_id: 4,
-                data: stream::RangeBuf::from(b"aaaaa", 0, false),
+                data: RangeBuf::from(b"aaaaa", 0, false),
             })
         );
     }
@@ -13358,7 +13359,7 @@ mod tests {
 
         let frames = [frame::Frame::Stream {
             stream_id: 0,
-            data: stream::RangeBuf::from(b"aa", 0, false),
+            data: RangeBuf::from(b"aa", 0, false),
         }];
 
         let pkt_type = packet::Type::Short;
@@ -13815,7 +13816,7 @@ mod tests {
             iter.next(),
             Some(&frame::Frame::Stream {
                 stream_id: 8,
-                data: stream::RangeBuf::from(b"aaaaaaaaaa", 0, false),
+                data: RangeBuf::from(b"aaaaaaaaaa", 0, false),
             })
         );
 
@@ -13863,7 +13864,7 @@ mod tests {
             iter.next(),
             Some(&frame::Frame::Stream {
                 stream_id: 0,
-                data: stream::RangeBuf::from(b"aaaaaaaaaaaaaaa", 0, false),
+                data: RangeBuf::from(b"aaaaaaaaaaaaaaa", 0, false),
             })
         );
 
@@ -13885,7 +13886,7 @@ mod tests {
             iter.next(),
             Some(&frame::Frame::Stream {
                 stream_id: 4,
-                data: stream::RangeBuf::from(b"a", 0, false),
+                data: RangeBuf::from(b"a", 0, false),
             })
         );
 
@@ -14523,7 +14524,7 @@ mod tests {
 
             assert_eq!(stream, &frame::Frame::Stream {
                 stream_id: 8,
-                data: stream::RangeBuf::from(&out, off, false),
+                data: RangeBuf::from(&out, off, false),
             });
 
             off = match stream {
@@ -14546,7 +14547,7 @@ mod tests {
 
             assert_eq!(stream, &frame::Frame::Stream {
                 stream_id: 16,
-                data: stream::RangeBuf::from(&out, off, false),
+                data: RangeBuf::from(&out, off, false),
             });
 
             off = match stream {
@@ -14569,7 +14570,7 @@ mod tests {
 
             assert_eq!(stream, &frame::Frame::Stream {
                 stream_id: 20,
-                data: stream::RangeBuf::from(&out, off, false),
+                data: RangeBuf::from(&out, off, false),
             });
 
             off = match stream {
@@ -14593,7 +14594,7 @@ mod tests {
                 frames.first(),
                 Some(&frame::Frame::Stream {
                     stream_id: 12,
-                    data: stream::RangeBuf::from(&out, off, false),
+                    data: RangeBuf::from(&out, off, false),
                 })
             );
 
@@ -14607,7 +14608,7 @@ mod tests {
 
             assert_eq!(stream, &frame::Frame::Stream {
                 stream_id: 4,
-                data: stream::RangeBuf::from(&out, off, false),
+                data: RangeBuf::from(&out, off, false),
             });
 
             off = match stream {
@@ -14630,7 +14631,7 @@ mod tests {
 
             assert_eq!(stream, &frame::Frame::Stream {
                 stream_id: 0,
-                data: stream::RangeBuf::from(&out, off, false),
+                data: RangeBuf::from(&out, off, false),
             });
 
             off = match stream {
@@ -14715,7 +14716,7 @@ mod tests {
             frames.first(),
             Some(&frame::Frame::Stream {
                 stream_id: 8,
-                data: stream::RangeBuf::from(b"b", 0, false),
+                data: RangeBuf::from(b"b", 0, false),
             })
         );
 
@@ -14729,7 +14730,7 @@ mod tests {
             frames.first(),
             Some(&frame::Frame::Stream {
                 stream_id: 0,
-                data: stream::RangeBuf::from(b"b", 0, false),
+                data: RangeBuf::from(b"b", 0, false),
             })
         );
 
@@ -14743,7 +14744,7 @@ mod tests {
             frames.first(),
             Some(&frame::Frame::Stream {
                 stream_id: 12,
-                data: stream::RangeBuf::from(b"b", 0, false),
+                data: RangeBuf::from(b"b", 0, false),
             })
         );
 
@@ -14756,7 +14757,7 @@ mod tests {
             frames.first(),
             Some(&frame::Frame::Stream {
                 stream_id: 4,
-                data: stream::RangeBuf::from(b"b", 0, false),
+                data: RangeBuf::from(b"b", 0, false),
             })
         );
 
@@ -14855,7 +14856,7 @@ mod tests {
 
             assert_eq!(stream, &frame::Frame::Stream {
                 stream_id: 0,
-                data: stream::RangeBuf::from(&out, off_0, false),
+                data: RangeBuf::from(&out, off_0, false),
             });
 
             off_0 = match stream {
@@ -14889,7 +14890,7 @@ mod tests {
 
             assert_eq!(stream, &frame::Frame::Stream {
                 stream_id: 4,
-                data: stream::RangeBuf::from(&out, off_4, false),
+                data: RangeBuf::from(&out, off_4, false),
             });
 
             off_4 = match stream {
@@ -14960,7 +14961,7 @@ mod tests {
             iter.next(),
             Some(&frame::Frame::Stream {
                 stream_id: 4,
-                data: stream::RangeBuf::from(b"b", 0, false),
+                data: RangeBuf::from(b"b", 0, false),
             })
         );
         assert_eq!(pipe.client.stats().retrans, 1);
@@ -18622,6 +18623,7 @@ mod packet;
 mod path;
 mod pmtud;
 mod rand;
+mod range_buf;
 mod ranges;
 mod recovery;
 mod stream;
