+++ quiche/src/range_buf.rs
@@ -1,0 +1,163 @@
+// Copyright (C) 2025, Cloudflare, Inc.
+// All rights reserved.
+//
+// Redistribution and use in source and binary forms, with or without
+// modification, are permitted provided that the following conditions are
+// met:
+//
+//     * Redistributions of source code must retain the above copyright notice,
+//       this list of conditions and the following disclaimer.
+//
+//     * Redistributions in binary form must reproduce the above copyright
+//       notice, this list of conditions and the following disclaimer in the
+//       documentation and/or other materials provided with the distribution.
+//
+// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS
+// IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO,
+// THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR
+// PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR
+// CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
+// EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
+// PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
+// PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
+// LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
+// NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+// SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+
+use std::cmp;
+use std::sync::Arc;
+
+/// Buffer holding data at a specific offset.
+///
+/// The data is stored in a `Vec<u8>` in such a way that it can be shared
+/// between multiple `RangeBuf` objects.
+///
+/// Each `RangeBuf` will have its own view of that buffer, where the `start`
+/// value indicates the initial offset within the `Vec`, and `len` indicates the
+/// number of bytes, starting from `start` that are included.
+///
+/// In addition, `pos` indicates the current offset within the `Vec`, starting
+/// from the very beginning of the `Vec`.
+///
+/// Finally, `off` is the starting offset for the specific `RangeBuf` within the
+/// stream the buffer belongs to.
+#[derive(Clone, Debug, Default, Eq)]
+pub struct RangeBuf {
+    /// The internal buffer holding the data.
+    ///
+    /// To avoid needless allocations when a RangeBuf is split, this field is
+    /// reference-counted and can be shared between multiple RangeBuf objects,
+    /// and sliced using the `start` and `len` values.
+    pub(crate) data: Arc<Vec<u8>>,
+
+    /// The initial offset within the internal buffer.
+    pub(crate) start: usize,
+
+    /// The current offset within the internal buffer.
+    pub(crate) pos: usize,
+
+    /// The number of bytes in the buffer, from the initial offset.
+    pub(crate) len: usize,
+
+    /// The offset of the buffer within a stream.
+    pub(crate) off: u64,
+
+    /// Whether this contains the final byte in the stream.
+    pub(crate) fin: bool,
+}
+
+impl RangeBuf {
+    /// Creates a new `RangeBuf` from the given slice.
+    pub fn from(buf: &[u8], off: u64, fin: bool) -> RangeBuf {
+        RangeBuf {
+            data: Arc::new(Vec::from(buf)),
+            start: 0,
+            pos: 0,
+            len: buf.len(),
+            off,
+            fin,
+        }
+    }
+
+    /// Returns whether `self` holds the final offset in the stream.
+    pub fn fin(&self) -> bool {
+        self.fin
+    }
+
+    /// Returns the starting offset of `self`.
+    pub fn off(&self) -> u64 {
+        (self.off - self.start as u64) + self.pos as u64
+    }
+
+    /// Returns the final offset of `self`.
+    pub fn max_off(&self) -> u64 {
+        self.off() + self.len() as u64
+    }
+
+    /// Returns the length of `self`.
+    pub fn len(&self) -> usize {
+        self.len - (self.pos - self.start)
+    }
+
+    /// Returns true if `self` has a length of zero bytes.
+    pub fn is_empty(&self) -> bool {
+        self.len() == 0
+    }
+
+    /// Consumes the starting `count` bytes of `self`.
+    pub fn consume(&mut self, count: usize) {
+        self.pos += count;
+    }
+
+    /// Splits the buffer into two at the given index.
+    pub fn split_off(&mut self, at: usize) -> RangeBuf {
+        assert!(
+            at <= self.len,
+            "`at` split index (is {}) should be <= len (is {})",
+            at,
+            self.len
+        );
+
+        let buf = RangeBuf {
+            data: self.data.clone(),
+            start: self.start + at,
+            pos: cmp::max(self.pos, self.start + at),
+            len: self.len - at,
+            off: self.off + at as u64,
+            fin: self.fin,
+        };
+
+        self.pos = cmp::min(self.pos, self.start + at);
+        self.len = at;
+        self.fin = false;
+
+        buf
+    }
+}
+
+impl std::ops::Deref for RangeBuf {
+    type Target = [u8];
+
+    fn deref(&self) -> &[u8] {
+        &self.data[self.pos..self.start + self.len]
+    }
+}
+
+impl Ord for RangeBuf {
+    fn cmp(&self, other: &RangeBuf) -> cmp::Ordering {
+        // Invert ordering to implement min-heap.
+        self.off.cmp(&other.off).reverse()
+    }
+}
+
+impl PartialOrd for RangeBuf {
+    fn partial_cmp(&self, other: &RangeBuf) -> Option<cmp::Ordering> {
+        Some(self.cmp(other))
+    }
+}
+
+impl PartialEq for RangeBuf {
+    fn eq(&self, other: &RangeBuf) -> bool {
+        self.off == other.off
+    }
+}
