+++ quiche/src/stream/mod.rs
@@ -893,143 +893,10 @@ impl ExactSizeIterator for StreamIter {
     }
 }
 
-/// Buffer holding data at a specific offset.
-///
-/// The data is stored in a `Vec<u8>` in such a way that it can be shared
-/// between multiple `RangeBuf` objects.
-///
-/// Each `RangeBuf` will have its own view of that buffer, where the `start`
-/// value indicates the initial offset within the `Vec`, and `len` indicates the
-/// number of bytes, starting from `start` that are included.
-///
-/// In addition, `pos` indicates the current offset within the `Vec`, starting
-/// from the very beginning of the `Vec`.
-///
-/// Finally, `off` is the starting offset for the specific `RangeBuf` within the
-/// stream the buffer belongs to.
-#[derive(Clone, Debug, Default, Eq)]
-pub struct RangeBuf {
-    /// The internal buffer holding the data.
-    ///
-    /// To avoid needless allocations when a RangeBuf is split, this field is
-    /// reference-counted and can be shared between multiple RangeBuf objects,
-    /// and sliced using the `start` and `len` values.
-    data: Arc<Vec<u8>>,
-
-    /// The initial offset within the internal buffer.
-    start: usize,
-
-    /// The current offset within the internal buffer.
-    pos: usize,
-
-    /// The number of bytes in the buffer, from the initial offset.
-    len: usize,
-
-    /// The offset of the buffer within a stream.
-    off: u64,
-
-    /// Whether this contains the final byte in the stream.
-    fin: bool,
-}
-
-impl RangeBuf {
-    /// Creates a new `RangeBuf` from the given slice.
-    pub fn from(buf: &[u8], off: u64, fin: bool) -> RangeBuf {
-        RangeBuf {
-            data: Arc::new(Vec::from(buf)),
-            start: 0,
-            pos: 0,
-            len: buf.len(),
-            off,
-            fin,
-        }
-    }
-
-    /// Returns whether `self` holds the final offset in the stream.
-    pub fn fin(&self) -> bool {
-        self.fin
-    }
-
-    /// Returns the starting offset of `self`.
-    pub fn off(&self) -> u64 {
-        (self.off - self.start as u64) + self.pos as u64
-    }
-
-    /// Returns the final offset of `self`.
-    pub fn max_off(&self) -> u64 {
-        self.off() + self.len() as u64
-    }
-
-    /// Returns the length of `self`.
-    pub fn len(&self) -> usize {
-        self.len - (self.pos - self.start)
-    }
-
-    /// Returns true if `self` has a length of zero bytes.
-    pub fn is_empty(&self) -> bool {
-        self.len() == 0
-    }
-
-    /// Consumes the starting `count` bytes of `self`.
-    pub fn consume(&mut self, count: usize) {
-        self.pos += count;
-    }
-
-    /// Splits the buffer into two at the given index.
-    pub fn split_off(&mut self, at: usize) -> RangeBuf {
-        assert!(
-            at <= self.len,
-            "`at` split index (is {}) should be <= len (is {})",
-            at,
-            self.len
-        );
-
-        let buf = RangeBuf {
-            data: self.data.clone(),
-            start: self.start + at,
-            pos: cmp::max(self.pos, self.start + at),
-            len: self.len - at,
-            off: self.off + at as u64,
-            fin: self.fin,
-        };
-
-        self.pos = cmp::min(self.pos, self.start + at);
-        self.len = at;
-        self.fin = false;
-
-        buf
-    }
-}
-
-impl std::ops::Deref for RangeBuf {
-    type Target = [u8];
-
-    fn deref(&self) -> &[u8] {
-        &self.data[self.pos..self.start + self.len]
-    }
-}
-
-impl Ord for RangeBuf {
-    fn cmp(&self, other: &RangeBuf) -> cmp::Ordering {
-        // Invert ordering to implement min-heap.
-        self.off.cmp(&other.off).reverse()
-    }
-}
-
-impl PartialOrd for RangeBuf {
-    fn partial_cmp(&self, other: &RangeBuf) -> Option<cmp::Ordering> {
-        Some(self.cmp(other))
-    }
-}
-
-impl PartialEq for RangeBuf {
-    fn eq(&self, other: &RangeBuf) -> bool {
-        self.off == other.off
-    }
-}
-
 #[cfg(test)]
 mod tests {
+    use crate::range_buf::RangeBuf;
+
     use super::*;
 
     #[test]
